Refactor Logging File into Library Module
