+++ src/mavedb/lib/logging/__init__.py
@@ -1,0 +1,9 @@
+import logging
+
+from mavedb.lib.logging.logged_route import LoggedRoute
+
+__all__ = [
+    "LoggedRoute",
+]
+
+logging.basicConfig(level=logging.INFO)
